new data reading layout
