+++ nt2/__init__.py
@@ -1,1 +1,4 @@
-__version__ = "0.4.1"
+__version__ = "0.5.0"
+
+from nt2.read import Data as Data
+from nt2.plotters import polarplot as polarplot
